front-end/requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.2
google-api-core==2.1.0
google-auth==2.3.0
google-cloud-compute==0.5.0
googleapis-common-protos==1.53.0
grpcio==1.41.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
packaging==21.0
proto-plus==1.19.4
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.19.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
